feat: Implementação da camada de Repository, Service e Controller, Implementação de Login

+++ src/main/java/br/upf/biblioteca/dto/AutorDTO.java
@@ -26,7 +26,7 @@ public class AutorDTO {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "CD_AUTOR")
-	private Integer cdAutor;
+	private Long cdAutor;
 	
 	@Column(name = "NM_NOME", nullable = false)
 	private String nmNome;
